+++ src/Fable.Cli/Contributors.fs
@@ -41,7 +41,7 @@ let getRandom() =
           "NickDarvey";     "thinkbeforecoding";  "cartermp";
           "chkn";           "MNie";               "Choc13";
           "davedawkins";    "njlr";               "steveofficer";
-          "cannorin"
+          "cannorin";       "hensou";
         |]
     Array.length contributors
     |> System.Random().Next